feat: refactor CommunityRatings component to use separate StarIcon component for better code organization
